+++ NEWS.md
@@ -1,3 +1,9 @@
+# UCSCXenaShiny 0.1.1
+
+* Used **testthat** package
+* Fixed path of `NEWS.md`
+* Fixed error: cannot find object 'XenaInfo'
+
 # UCSCXenaShiny 0.1.0
 
 * Submitted to CRAN